+++ R/class_runtime.R
@@ -1,24 +1,28 @@
 runtime_init <- function(
   target = NULL,
   frames = NULL,
-  interactive = NULL
+  interactive = NULL,
+  store = NULL
 ) {
   runtime_new(
     target = target,
     frames = frames,
-    interactive = interactive
+    interactive = interactive,
+    store = store
   )
 }
 
 runtime_new <- function(
   target = NULL,
   frames = NULL,
-  interactive = NULL
+  interactive = NULL,
+  store = NULL
 ) {
   runtime_class$new(
     target = target,
     frames = frames,
-    interactive = interactive
+    interactive = interactive,
+    store = store
   )
 }
 
@@ -31,14 +35,17 @@ runtime_class <- R6::R6Class(
     target = NULL,
     frames = NULL,
     interactive = NULL,
+    store = NULL,
     initialize = function(
       target = NULL,
       frames = NULL,
-      interactive = NULL
+      interactive = NULL,
+      store = NULL
     ) {
       self$target <- target
       self$frames <- frames
       self$interactive <- interactive
+      self$store <- store
     },
     exists_target = function() {
       !is.null(self$target)
@@ -49,6 +56,9 @@ runtime_class <- R6::R6Class(
     exists_interactive = function() {
       !is.null(self$interactive)
     },
+    exists_store = function() {
+      !is.null(self$store)
+    },
     get_target = function() {
       self$target
     },
@@ -57,6 +67,9 @@ runtime_class <- R6::R6Class(
     },
     get_interactive = function() {
       self$interactive
+    },
+    get_store = function() {
+      self$store
     },
     set_target = function(target) {
       self$target <- target
@@ -67,6 +80,9 @@ runtime_class <- R6::R6Class(
     set_interactive = function(interactive) {
       self$interactive <- interactive
     },
+    set_store = function(store) {
+      self$store <- store
+    },
     unset_target = function() {
       self$target <- NULL
     },
@@ -75,6 +91,9 @@ runtime_class <- R6::R6Class(
     },
     unset_interactive = function() {
       self$interactive <- NULL
+    },
+    unset_store = function() {
+      self$store <- NULL
     },
     validate = function() {
       if (!is.null(self$target)) {
@@ -87,6 +106,10 @@ runtime_class <- R6::R6Class(
       if (!is.null(self$interactive)) {
         tar_assert_scalar(self$interactive)
         tar_assert_lgl(self$interactive)
+      }
+      if (!is.null(self$store)) {
+        tar_assert_scalar(self$store)
+        tar_assert_chr(self$store)
       }
     }
   )