+++ R/tar_path.R
@@ -24,7 +24,6 @@
 #'   upload it to the AWS S3 bucket. This does not apply to formats
 #'   `"file"` or `"aws_file"`, where you would never need `storage = "none"`
 #'   anyway.
-#' @inheritParams tar_validate
 #' @param name Symbol, name of a target.
 #'   If `NULL`, `tar_path()` returns the path of the target currently running
 #'   in a pipeline.
@@ -37,6 +36,11 @@
 #'   This is useful if you are writing to `tar_path()` from inside a
 #'   `storage = "none"` target and need the parent directory of the file
 #'   to exist.
+#' @param store Character of length 1, path to the data store if `tar_path()`
+#'   is called outside a running pipeline. If `tar_path()` is called
+#'   inside a running pipeline, this argument is ignored
+#'   and actual the path to the running pipeline's data store
+#'   is used instead.
 #' @examples
 #' tar_path()
 #' tar_path(your_target)
@@ -75,7 +79,7 @@ tar_path_running <- function(default, path_store) {
     store_tar_path(
       tar_runtime$get_target()$store,
       tar_runtime$get_target(),
-      path_store
+      tar_runtime$get_store()
     ),
     as.character(default)
   )